Move all IDE script templates to `IdeScriptTemplates.kt`
And restore `KotlinSettingsScript.kt` for binary compatibility.
