+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/SettingsScriptApi.kt
@@ -24,21 +24,14 @@ import org.gradle.api.file.CopySpec
 import org.gradle.api.file.DeleteSpec
 import org.gradle.api.file.FileTree
 import org.gradle.api.initialization.Settings
-import org.gradle.api.internal.GradleInternal
 import org.gradle.api.internal.ProcessOperations
-import org.gradle.api.internal.file.DefaultFileOperations
-import org.gradle.api.internal.file.FileCollectionFactory
-import org.gradle.api.internal.file.FileLookup
 import org.gradle.api.internal.file.FileOperations
-import org.gradle.api.invocation.Gradle
 import org.gradle.api.logging.Logger
 import org.gradle.api.logging.Logging
 import org.gradle.api.logging.LoggingManager
 import org.gradle.api.resources.ResourceHandler
 import org.gradle.api.tasks.WorkResult
-import org.gradle.internal.service.ServiceRegistry
 import org.gradle.kotlin.dsl.support.delegates.SettingsDelegate
-import org.gradle.kotlin.dsl.support.get
 import org.gradle.kotlin.dsl.support.internalError
 import org.gradle.kotlin.dsl.support.serviceOf
 import org.gradle.kotlin.dsl.support.unsafeLazy
@@ -374,7 +367,7 @@ abstract class SettingsScriptApi(
     @Incubating
     @Suppress("unused")
     open fun pluginManagement(@Suppress("unused_parameter") block: PluginManagementSpec.() -> Unit): Unit =
-            internalError()
+        internalError()
 
     /**
      * Configures the build script classpath for settings.
@@ -383,28 +376,5 @@ abstract class SettingsScriptApi(
      */
     @Suppress("unused")
     open fun buildscript(@Suppress("unused_parameter") block: ScriptHandlerScope.() -> Unit): Unit =
-            internalError()
-}
-
-
-internal
-fun fileOperationsFor(settings: Settings): FileOperations =
-    fileOperationsFor(settings.gradle, settings.rootDir)
-
-
-internal
-fun fileOperationsFor(gradle: Gradle, baseDir: File?): FileOperations =
-    fileOperationsFor((gradle as GradleInternal).services, baseDir)
-
-
-internal
-fun fileOperationsFor(services: ServiceRegistry, baseDir: File?): FileOperations {
-    val fileLookup = services.get<FileLookup>()
-    val fileResolver = baseDir?.let { fileLookup.getFileResolver(it) } ?: fileLookup.fileResolver
-    val fileCollectionFactory = services.get<FileCollectionFactory>().withResolver(fileResolver)
-    return DefaultFileOperations.createSimple(
-        fileResolver,
-        fileCollectionFactory,
-        services
-    )
+        internalError()
 }
